fix(docker): Upgrade node version to resolve npm incompatibility
This commit upgrades the Node.js version in the Dockerfile from 'node:18-alpine' to 'node:20-alpine'. This change is necessary to resolve an 'EBADENGINE' error that occurred during the Docker build, where the latest version of npm was not compatible with the older Node.js version.

By upgrading the Node.js image, we ensure compatibility with the latest npm, allowing the Docker build to complete successfully.